requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
jsonschema==4.23.0
cryptography==43.0.1
python-dotenv==1.0.1
httpx==0.27.2
